taking care of CR comments
