+++ config/exchanges/IN-EA_IN-WE.yaml
@@ -2,5 +2,5 @@ lonlat:
   - 83.363
   - 21.456
 parsers:
-  exchange: IN_WE_new.fetch_exchange
+  exchange: IN_WE.fetch_exchange
 rotation: 310
