+++ config/exchanges/IN-NO_IN-WE.yaml
@@ -2,5 +2,5 @@ lonlat:
   - 75.065
   - 24.892
 parsers:
-  exchange: IN_WE_new.fetch_exchange
+  exchange: IN_WE.fetch_exchange
 rotation: 150
