+++ config/exchanges/IN-SO_IN-WE.yaml
@@ -3,5 +3,5 @@ lonlat:
   - 17.702
 
 parsers:
-  exchange: IN_WE_new.fetch_exchange
+  exchange: IN_WE.fetch_exchange
 rotation: -30
